Prevented GitDependencies.sh to link /usr/lib/libc.dylib if the file doesn't exist. That linked file doesn't exist anymore on Mac OS Big Sur (but does on Mac OS Catalina).   - If the Setup.command (invoking GitDependencies.sh) is ran more than once, the first time the link (to a non existing file) is created, but any other times, the file test followed the link and said the file didn't exist, then the script tried to create the symbolic link again, but this time the link existed (just pointing to a non-existing file) and then the 'ln' command failed and the shell script unexpectdly aborted.
#jira UE-131354 - Creating a Blank BP Project with Starter Content Loads Into an Empty Level
#rb Ben.March, Will.Damon
#fyi Ryan.Hummer
#lockdown mitchell.wilson

[CL 17873078 by Patrick Laflamme in 4.27 branch]

+++ Engine/Build/BatchFiles/Mac/GitDependencies.sh
@@ -19,7 +19,7 @@ source SetupEnvironment.sh -mono "`pwd`"
 
 cd ../../../..
 
-if [ ! -f Engine/Binaries/ThirdParty/Mono/Mac/lib/libmsvcrt.dylib ]; then
+if [ ! -f Engine/Binaries/ThirdParty/Mono/Mac/lib/libmsvcrt.dylib ] && [ -f /usr/lib/libc.dylib ]; then
 	ln -s /usr/lib/libc.dylib Engine/Binaries/ThirdParty/Mono/Mac/lib/libmsvcrt.dylib
 fi
